requirements.txt
pytest
webdriver-manager==3.5.2
requests
selenium
urllib3
faker
allure-pytest
flake8